Use single curly brace according to fzf --preview documentation
The fzf documentation says that to use the output of each line inside
the preview command, the placeholders should be of the form `{[0-9]+}`.

> Also, \fB{q}\fR is replaced to the current query string, and \fB{n}\fR
> is replaced to zero-based ordinal index of the line. Use \fB{+n}\fR if
> you want all index numbers when multiple lines are selected.
> -- https://github.com/junegunn/fzf/blob/master/man/man1/fzf.1#L363-L365

https://github.com/junegunn/fzf

The preview commands generated by git-fzf right now use `{{[0-9]+}}`
(note the wrapping with the extra curly brace).

I am not able to use `{{1}}` to preview with fzf version 0.17.5 or fzf
version 0.18. I can't find any documentation about whether this was a
breaking change inside of fzf, I can't find any reference to `{{.}}`
anywhere in the [changelog][1].

[1]: https://github.com/junegunn/fzf/blob/master/CHANGELOG.md